Fix references to generic types
Also refactor most code into modules.
